requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
marshmallow==3.21.3
marshmallow-sqlalchemy==1.1.0
python-dotenv==1.0.1
Flask-JWT-Extended==4.6.0
flasgger==0.9.7.1
